Test #2 (override location)
I was unable to demo this feature at review, and it looked like
I'd broken it while implementing #3.  On further triage, it seems
that the feature does in fact work, but as it not longer had any
tests, I've reinstated them here, to make sure this case is
explicitly tested.