Refactor to allow controller subclasses
Progresses #8.  Oddly, the Controllers are simple subclasses, e.g.
PC::X, whereas the Views are using plugins, e.g. PC::Plugin::Y.

This seems to make sense at the moment, because of how OX is structured:
e.g. I want a single View class, because that's then easier to override
for testing, but the same restriction doesn't apply for Controller, so
why bother with the additional complexity of plugins?  Not sure if I've
come to correct decision on this, but will see how it goes for now.