Refactor rendering into P::RouteBuilder
All controllers now have to simply implement a method for 'get', 'post',
etc., or 'any' -- exactly like OX::RouteBuilder::HTTPMethod.  However,
instead of returning an OX::Response, they just return data (as Index
already did from the start, but which was more fiddly with the get/post
handlers of NewEvent) .  The RouteBuilder takes care of rendering the
HTML and returning a Response.