Benchmarks TT vs Mojo::DOM
As perlancar asked about it in:
http://osfameron.github.io/pageboy/posts/2015/10/25/accidental-templating-system.html#comments
and I can't sleep...

         Rate   mojo    tt
    mojo 38.9/s   --  -19%
    tt   47.8/s  23%    --

So the Mojo::DOM approach *is* slower, as I expected, but really,
that's not bad considering TT is pretty well optimized for speed.
Looking at the templates compared to the animated HTML ones, I'm
fairly happy with the current approach, but obviously if your profiling
indicates that templating is a major bottleneck, then this might
be an issue.